Fixed Directory Issues
Changed the file paths in the C++ code to relative file paths as per
link 1.
Changed the file path in Python code so it looks for the .dat files in
.../Current_Data.
User only need change the "use custom working directory" to the
project's directory in the options tab of the schemes for the file
paths to work on any Mac machine.

http://www.gidforums.com/t-19393.html